Moved rotation handle code to vispy, less redundancy and handle length constant
